+++ src/aiofiles/__init__.py
@@ -1,4 +1,5 @@
 """Utilities for asyncio-friendly file handling."""
+
 from .threadpool import (
     open,
     stdin,
