+++ src/aiofiles/ospath.py
@@ -1,30 +1,37 @@
 """Async executor versions of file functions from the os.path module."""
 
-import asyncio
-from functools import partial, wraps
 from os import path
 
+from .base import wrap
 
-def wrap(func):
-    @wraps(func)
-    async def run(*args, loop=None, executor=None, **kwargs):
-        if loop is None:
-            loop = asyncio.get_running_loop()
-        pfunc = partial(func, *args, **kwargs)
-        return await loop.run_in_executor(executor, pfunc)
+__all__ = [
+    "abspath",
+    "getatime",
+    "getctime",
+    "getmtime",
+    "getsize",
+    "exists",
+    "isdir",
+    "isfile",
+    "islink",
+    "ismount",
+    "samefile",
+    "sameopenfile",
+]
 
-    return run
+abspath = wrap(path.abspath)
 
-
-exists = wrap(path.exists)
-isfile = wrap(path.isfile)
-isdir = wrap(path.isdir)
-islink = wrap(path.islink)
-ismount = wrap(path.ismount)
-getsize = wrap(path.getsize)
-getmtime = wrap(path.getmtime)
 getatime = wrap(path.getatime)
 getctime = wrap(path.getctime)
+getmtime = wrap(path.getmtime)
+getsize = wrap(path.getsize)
+
+exists = wrap(path.exists)
+
+isdir = wrap(path.isdir)
+isfile = wrap(path.isfile)
+islink = wrap(path.islink)
+ismount = wrap(path.ismount)
+
 samefile = wrap(path.samefile)
 sameopenfile = wrap(path.sameopenfile)
-abspath = wrap(path.abspath)
